+++ .gitignore
@@ -1,5 +1,6 @@
 /.gradle
 /.idea
+/logs
 **/build
 *.log
 *.class
