+++ docker/sge/docker-compose.yaml
@@ -2,6 +2,8 @@ version: "3.8"
 services:
 
   sge-container:
+    environment:
+      TARGET_GRID_ENGINE: SGE
     build:
       context: ../..
       dockerfile: ./docker/sge/Dockerfile
