+++ src/main/java/com/epam/grid/engine/mapper/host/slurm/SlurmHostMapper.java
@@ -5,8 +5,9 @@ import com.epam.grid.engine.entity.host.slurm.SlurmHost;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.Mappings;
+import org.mapstruct.ReportingPolicy;
 
-@Mapper(componentModel = "spring")
+@Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface SlurmHostMapper {
 
     @Mappings({
