+++ src/main/resources/application.properties
@@ -19,7 +19,7 @@
 
 # Common properties
 server.port=8080
-grid.engine.type=SLURM
+grid.engine.type=${TARGET_GRID_ENGINE:SGE}
 spring.jackson.default-property-inclusion = NON_NULL
 command.template.path=templates/
 job.log.dir=${GE_JOB_LOGS:logs}/
@@ -32,4 +32,3 @@ sge.execd.port=6445
 sge.qmaster.host.path=/opt/sge/default/common/act_qmaster
 sge.parallel.environment.registration.default.slots=${GRID_ENGINE_API_PE_REGISTRATION_DEFAULT_SLOTS:999}
 sge.parallel.environment.registration.default.allocation.rule=${GRID_ENGINE_API_PE_REG_DEFAULT_ALLOCATION_RULE:$fill_up}
-
