+++ src/main/resources/templates/slurm/command/scontrol
@@ -1,5 +1,5 @@
-[# th:if="${filter != null and not #lists.isEmpty(filter.hosts)}"]
 scontrol -o show node
+[# th:if="${filter != null and not #lists.isEmpty(filter.hosts)}"]
 [# th:text="${#strings.listJoin(filter.hosts, ',')}"]
  [/]
 [/]
